token added, login and logout

+++ app/services/UserService.scala
@@ -8,12 +8,28 @@ import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
 class UserService @Inject()(userDAO: UserDAO)(implicit ec: ExecutionContext) {
+  private val ADMIN: String = "Admin"
 
   def createUser(user: User): Future[User] = {
+    if (user.role.equalsIgnoreCase(ADMIN)) {
+      checkForAdminUser().flatMap {
+        case Some(_) => Future.failed(new Exception("Failed to create admin user, Admin User already exists"))
+        case None => handleCreate(user)
+      }
+    } else {
+      handleCreate(user)
+    }
+  }
+
+  private def handleCreate(user: User): Future[User] = {
     userDAO.create(user).recoverWith {
       case e: Exception =>
         Future.failed(new Exception("Failed to create user", e))
     }
+  }
+
+  private def checkForAdminUser(): Future[Option[User]] = {
+    userDAO.fetchAdminUser()
   }
 
   def updateUser(id: Int, name: String, email: String, password: String, phoneNumber: String, role: String, token: Option[String]): Future[Int] = {
@@ -25,9 +41,16 @@ class UserService @Inject()(userDAO: UserDAO)(implicit ec: ExecutionContext) {
   }
 
   def deleteUser(id: Int): Future[Int] = {
-    userDAO.delete(id).recoverWith {
-      case e: Exception =>
-        Future.failed(new Exception(s"Failed to delete user with ID $id", e))
+    getUser(id).flatMap {
+      case Some(user) =>
+        if(user.role.equalsIgnoreCase(ADMIN)) {
+         return Future.failed(new Exception("Admin User can not be deleted."))
+        }
+        userDAO.delete(id).recoverWith {
+          case e: Exception =>
+            Future.failed(new Exception(s"Failed to delete user with ID $id", e))
+        }
+      case _ => Future.failed(new Exception("User does not exist"))
     }
   }
 
@@ -45,6 +68,13 @@ class UserService @Inject()(userDAO: UserDAO)(implicit ec: ExecutionContext) {
     }
   }
 
+  def getUserByToken(token: String): Future[Option[User]] = {
+    userDAO.findByToken(token).recoverWith {
+      case e: Exception =>
+        Future.failed(new Exception("Failed to find user "))
+    }
+  }
+
   def listUsers(): Future[Seq[User]] = {
     userDAO.findAll().recoverWith {
       case e: Exception =>
@@ -55,6 +85,7 @@ class UserService @Inject()(userDAO: UserDAO)(implicit ec: ExecutionContext) {
   def login(user: User): Future[String] = {
     userDAO.login(user, generateRandomString())
   }
+
   def logout(user: User): Unit = {
     userDAO.logout(user)
   }
